Fix `__getDepthMap()` for LibSass
The `__getDepthMap()` function makes LibSass hang while compiling. That's because the `index()` function doesn't seem to work exactly the same in LibSass (up to 3.4.3) as in Ruby Sass (up to 3.4.23), especially while looking for values that contain spaces inside comma-separated lists. Re-parsing the `$current` list before trying to obtain the `$parentIndex` fixes the issue.